[IMP] sale_stock_available_to_promise_release_dropshipping: add test
